requirements.txt
streamlit
torch
transformers
numpy
opencv-python
Pillow
open-clip-torch
google-generativeai
altair
pandas
